populate writing context with helpful information for determining cell style

+++ src/main/java/com/mihai/writer/style/color/CellColor.java
@@ -1,21 +1,19 @@
 package com.mihai.writer.style.color;
 
-import com.mihai.writer.style.font.CellFont;
-
 import java.util.Objects;
 
 public class CellColor {
 
-    public static final CellColor BLACK = new CellColor((byte) 0, (byte) 0, (byte) 0);
+    public static final CellColor BLACK = new CellColor(0, 0, 0);
 
     private final byte red;
     private final byte green;
     private final byte blue;
 
-    public CellColor(byte red, byte green, byte blue) {
-        this.red = red;
-        this.green = green;
-        this.blue = blue;
+    public CellColor(int red, int green, int blue) {
+        this.red = (byte) red;
+        this.green = (byte) green;
+        this.blue = (byte) blue;
     }
 
     public byte getRed() {
